src/research_generation/research_ecosystem_engine.py
import numpy as np
import networkx as nx
import pandas as pd
from typing import Dict, List, Any, Optional, Tuple
from dataclasses import dataclass, field
import scipy.stats as stats

from src.knowledge_graph.distributed_knowledge_graph import DistributedKnowledgeGraphEngine
from src.research_generation.research_gap_engine import ResearchGapEngine
from src.semantic_understanding.semantic_intent_engine import SemanticIntentEngine

@dataclass
class ResearchDomain:
    """
    Represents a research domain with its evolution characteristics
    """
    name: str
    complexity: float
    innovation_rate: float
    interdisciplinary_potential: float
    knowledge_graph_density: float
    temporal_dynamics: Dict[str, Any] = field(default_factory=dict)

class ResearchEcosystemEngine:
    def __init__(
        self, 
        knowledge_graph: Optional[DistributedKnowledgeGraphEngine] = None,
        research_gap_engine: Optional[ResearchGapEngine] = None,
        semantic_engine: Optional[SemanticIntentEngine] = None
    ):
        """
        Initialize Research Ecosystem Modeling Engine
        
        Args:
            knowledge_graph: Distributed knowledge graph engine
            research_gap_engine: Research gap identification engine
            semantic_engine: Semantic understanding engine
        """
        # Initialize dependencies
        self.knowledge_graph = knowledge_graph or DistributedKnowledgeGraphEngine()
        self.research_gap_engine = research_gap_engine or ResearchGapEngine()
        self.semantic_engine = semantic_engine or SemanticIntentEngine()
        
        # Research domain configuration
        self.research_domains = [
            'technology', 'science', 'engineering', 
            'mathematics', 'computing', 'interdisciplinary'
        ]
        
        # Ecosystem modeling parameters
        self.evolution_parameters = {
            'innovation_decay_rate': 0.1,
            'complexity_growth_factor': 0.05,
            'interdisciplinary_threshold': 0.6
        }
    
    def model_research_ecosystem(
        self, 
        time_horizon: int = 5,
        granularity: str = 'yearly'
    ) -> List[ResearchDomain]:
        """
        Model the evolution of research domains over time
        
        Args:
            time_horizon: Number of time periods to model
            granularity: Time granularity ('yearly', 'quarterly')
        
        Returns:
            Projected research domain evolution
        """
        # Initialize research domains
        research_domains = self._initialize_research_domains()
        
        # Simulate domain evolution
        for period in range(1, time_horizon + 1):
            research_domains = self._evolve_research_domains(
                research_domains, 
                period, 
                granularity
            )
        
        return research_domains
    
    def _initialize_research_domains(self) -> List[ResearchDomain]:
        """
        Initialize research domains with baseline characteristics
        
        Returns:
            List of initialized research domains
        """
        initialized_domains = []
        
        for domain_name in self.research_domains:
            # Extract domain-specific knowledge graph
            domain_graph = self.knowledge_graph.extract_domain_subgraph(domain_name)
            
            # Compute domain characteristics
            domain = ResearchDomain(
                name=domain_name,
                complexity=self._compute_domain_complexity(domain_graph),
                innovation_rate=self._compute_innovation_rate(domain_graph),
                interdisciplinary_potential=self._assess_interdisciplinary_potential(domain_graph),
                knowledge_graph_density=nx.density(domain_graph),
                temporal_dynamics={
                    'initial_state': {
                        'timestamp': 0,
                        'key_entities': list(domain_graph.nodes)[:10]
                    }
                }
            )
            
            initialized_domains.append(domain)
        
        return initialized_domains
    
    def _evolve_research_domains(
        self, 
        domains: List[ResearchDomain], 
        period: int,
        granularity: str
    ) -> List[ResearchDomain]:
        """
        Simulate evolution of research domains
        
        Args:
            domains: Current research domains
            period: Current time period
            granularity: Time granularity
        
        Returns:
            Evolved research domains
        """
        evolved_domains = []
        
        for domain in domains:
            # Compute domain evolution parameters
            innovation_decay = self._compute_innovation_decay(domain)
            complexity_growth = self._compute_complexity_growth(domain)
            
            # Update domain characteristics
            evolved_domain = ResearchDomain(
                name=domain.name,
                complexity=domain.complexity * (1 + complexity_growth),
                innovation_rate=max(0, domain.innovation_rate * (1 - innovation_decay)),
                interdisciplinary_potential=self._update_interdisciplinary_potential(domain),
                knowledge_graph_density=self._update_knowledge_graph_density(domain),
                temporal_dynamics=self._update_temporal_dynamics(domain, period, granularity)
            )
            
            evolved_domains.append(evolved_domain)
        
        return evolved_domains
    
    def _compute_domain_complexity(self, domain_graph: nx.DiGraph) -> float:
        """
        Compute domain complexity based on graph structure
        
        Args:
            domain_graph: Domain-specific knowledge graph
        
        Returns:
            Domain complexity score
        """
        # Compute complexity using graph metrics
        centrality_metrics = [
            nx.betweenness_centrality(domain_graph),
            nx.closeness_centrality(domain_graph),
            nx.pagerank(domain_graph)
        ]
        
        # Aggregate complexity metrics
        complexity_score = np.mean([
            np.mean(list(metric.values())) 
            for metric in centrality_metrics
        ])
        
        return complexity_score
    
    def _compute_innovation_rate(self, domain_graph: nx.DiGraph) -> float:
        """
        Compute domain innovation rate
        
        Args:
            domain_graph: Domain-specific knowledge graph
        
        Returns:
            Innovation rate score
        """
        # Compute innovation based on graph connectivity and node diversity
        node_diversity = len(domain_graph.nodes)
        edge_connectivity = nx.average_clustering(domain_graph)
        
        # Normalize and combine metrics
        innovation_score = (
            0.6 * edge_connectivity + 
            0.4 * (node_diversity / (node_diversity + 100))
        )
        
        return innovation_score
    
    def _assess_interdisciplinary_potential(self, domain_graph: nx.DiGraph) -> float:
        """
        Assess interdisciplinary potential of a domain
        
        Args:
            domain_graph: Domain-specific knowledge graph
        
        Returns:
            Interdisciplinary potential score
        """
        # Compute cross-domain connectivity
        external_connections = sum(
            1 for node in domain_graph.nodes 
            if any('cross-domain' in str(attr).lower() for attr in domain_graph.nodes[node].values())
        )
        
        # Compute interdisciplinary score
        interdisciplinary_score = external_connections / len(domain_graph.nodes)
        
        return interdisciplinary_score
    
    def _compute_innovation_decay(self, domain: ResearchDomain) -> float:
        """
        Compute innovation decay rate
        
        Args:
            domain: Research domain
        
        Returns:
            Innovation decay rate
        """
        # Decay based on current innovation rate and complexity
        decay_rate = (
            self.evolution_parameters['innovation_decay_rate'] * 
            (1 - domain.innovation_rate) * 
            (domain.complexity / 10)
        )
        
        return decay_rate
    
    def _compute_complexity_growth(self, domain: ResearchDomain) -> float:
        """
        Compute domain complexity growth
        
        Args:
            domain: Research domain
        
        Returns:
            Complexity growth factor
        """
        # Complexity growth based on innovation and interdisciplinary potential
        complexity_growth = (
            self.evolution_parameters['complexity_growth_factor'] * 
            domain.innovation_rate * 
            domain.interdisciplinary_potential
        )
        
        return complexity_growth
    
    def _update_interdisciplinary_potential(self, domain: ResearchDomain) -> float:
        """
        Update interdisciplinary potential over time
        
        Args:
            domain: Research domain
        
        Returns:
            Updated interdisciplinary potential
        """
        # Probabilistic increase in interdisciplinary potential
        if domain.interdisciplinary_potential < self.evolution_parameters['interdisciplinary_threshold']:
            # Increase potential with some randomness
            potential_increase = np.random.uniform(0.05, 0.15)
            return min(1.0, domain.interdisciplinary_potential + potential_increase)
        
        return domain.interdisciplinary_potential
    
    def _update_knowledge_graph_density(self, domain: ResearchDomain) -> float:
        """
        Update knowledge graph density
        
        Args:
            domain: Research domain
        
        Returns:
            Updated knowledge graph density
        """
        # Probabilistic density increase
        density_change = np.random.normal(0.05, 0.02)
        updated_density = domain.knowledge_graph_density * (1 + density_change)
        
        return max(0.0, min(1.0, updated_density))
    
    def _update_temporal_dynamics(
        self, 
        domain: ResearchDomain, 
        period: int, 
        granularity: str
    ) -> Dict[str, Any]:
        """
        Update temporal dynamics of a research domain
        
        Args:
            domain: Research domain
            period: Current time period
            granularity: Time granularity
        
        Returns:
            Updated temporal dynamics
        """
        # Copy existing temporal dynamics
        temporal_dynamics = domain.temporal_dynamics.copy()
        
        # Add new temporal state
        temporal_dynamics[f'{granularity}_{period}'] = {
            'timestamp': period,
            'complexity': domain.complexity,
            'innovation_rate': domain.innovation_rate,
            'interdisciplinary_potential': domain.interdisciplinary_potential
        }
        
        return temporal_dynamics
    
    def generate_research_trajectory_report(
        self, 
        domains: List[ResearchDomain]
    ) -> Dict[str, Any]:
        """
        Generate comprehensive research trajectory report
        
        Args:
            domains: Evolved research domains
        
        Returns:
            Research trajectory report
        """
        report = {
            'overall_ecosystem_metrics': self._compute_ecosystem_metrics(domains),
            'domain_evolution_details': {
                domain.name: self._generate_domain_report(domain)
                for domain in domains
            },
            'interdisciplinary_trends': self._analyze_interdisciplinary_trends(domains),
            'research_strategy_recommendations': self._generate_research_recommendations(domains)
        }
        
        return report
    
    def _compute_ecosystem_metrics(self, domains: List[ResearchDomain]) -> Dict[str, float]:
        """
        Compute overall research ecosystem metrics
        
        Args:
            domains: Research domains
        
        Returns:
            Ecosystem-level metrics
        """
        return {
            'average_complexity': np.mean([d.complexity for d in domains]),
            'average_innovation_rate': np.mean([d.innovation_rate for d in domains]),
            'average_interdisciplinary_potential': np.mean([d.interdisciplinary_potential for d in domains]),
            'ecosystem_diversity': len(domains)
        }
    
    def _generate_domain_report(self, domain: ResearchDomain) -> Dict[str, Any]:
        """
        Generate detailed report for a specific research domain
        
        Args:
            domain: Research domain
        
        Returns:
            Domain-specific research report
        """
        return {
            'name': domain.name,
            'complexity_trajectory': self._extract_temporal_trajectory(domain, 'complexity'),
            'innovation_trajectory': self._extract_temporal_trajectory(domain, 'innovation_rate'),
            'interdisciplinary_trajectory': self._extract_temporal_trajectory(domain, 'interdisciplinary_potential')
        }
    
    def _extract_temporal_trajectory(
        self, 
        domain: ResearchDomain, 
        metric: str
    ) -> List[float]:
        """
        Extract temporal trajectory for a specific metric
        
        Args:
            domain: Research domain
            metric: Metric to extract
        
        Returns:
            Temporal trajectory of the metric
        """
        trajectory = []
        
        for key, state in domain.temporal_dynamics.items():
            if isinstance(state, dict) and metric in state:
                trajectory.append(state[metric])
        
        return trajectory
    
    def _analyze_interdisciplinary_trends(self, domains: List[ResearchDomain]) -> Dict[str, Any]:
        """
        Analyze interdisciplinary trends across research domains
        
        Args:
            domains: Research domains
        
        Returns:
            Interdisciplinary trend analysis
        """
        return {
            'cross_domain_potential': np.mean([d.interdisciplinary_potential for d in domains]),
            'high_potential_domains': [
                d.name for d in domains 
                if d.interdisciplinary_potential > self.evolution_parameters['interdisciplinary_threshold']
            ]
        }
    
    def _generate_research_recommendations(self, domains: List[ResearchDomain]) -> List[Dict[str, Any]]:
        """
        Generate research strategy recommendations
        
        Args:
            domains: Research domains
        
        Returns:
            Research strategy recommendations
        """
        recommendations = []
        
        for domain in domains:
            if domain.interdisciplinary_potential > self.evolution_parameters['interdisciplinary_threshold']:
                recommendations.append({
                    'domain': domain.name,
                    'recommendation': f"Prioritize cross-domain collaboration in {domain.name}",
                    'rationale': f"High interdisciplinary potential of {domain.interdisciplinary_potential:.2f}"
                })
            
            if domain.innovation_rate < 0.3:
                recommendations.append({
                    'domain': domain.name,
                    'recommendation': f"Invest in innovation strategies for {domain.name}",
                    'rationale': f"Low innovation rate of {domain.innovation_rate:.2f}"
                })
        
        return recommendations
